# Conflicts:
#	demo/src/main/java/com/example/demo/DemoApplication.java

+++ demo/src/main/java/com/example/demo/Dao/productMapper.java
@@ -39,5 +39,5 @@ public interface productMapper {
 
 
     @Delete("delete from product where productNum=#{productNum}")
-    public void deleteProduct(int productNum);
+    public void deleteProduct(long productNum);
 }
